# web/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "The Unicorns farm API",
        "description": "This is a unicorn farm.\n     *                  No authorization is required to use this API.\n     *                  This documentation was generated by swagger. You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "contact": {
            "email": "support@example.com"
        },
        "version": "3.0.0"
    },
    "paths": {
        "/post/create": {
            "post": {
                "tags": [
                    "post"
                ],
                "summary": "Create a post.",
                "description": "Create a post",
                "operationId": "createPostAction",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "firstName",
                        "in": "query",
                        "description": "The user first name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "lastName",
                        "in": "query",
                        "description": "The user last name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "text",
                        "in": "query",
                        "description": "The content of the post ...",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "unicornId",
                        "in": "query",
                        "description": "The unicorn id to link the post to",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "examples": {
                            "application/json": {
                                "id": 1,
                                "user": {
                                    "id": 1,
                                    "firstName": "David",
                                    "lastName": "",
                                    "email": ""
                                },
                                "unicorn": null,
                                "text": "content of the post ...",
                                "createdAt": {
                                    "timezone": {
                                        "name": "Europe/Brussels",
                                        "location": {
                                            "country_code": "BE",
                                            "latitude": 50.83333,
                                            "longitude": 4.33333,
                                            "comments": ""
                                        }
                                    },
                                    "offset": 3600,
                                    "timestamp": 1548204515
                                },
                                "modifiedAt": {
                                    "timezone": {
                                        "name": "Europe/Brussels",
                                        "location": {
                                            "country_code": "BE",
                                            "latitude": 50.83333,
                                            "longitude": 4.33333,
                                            "comments": ""
                                        }
                                    },
                                    "offset": 3600,
                                    "timestamp": 1548204515
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post/link": {
            "post": {
                "tags": [
                    "post"
                ],
                "summary": "Link an existing post to a unicorn.",
                "description": "Link an existing post to a unicorn",
                "operationId": "linkPostToUnicornAction",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "postId",
                        "in": "query",
                        "description": "The id of the post",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "unicornId",
                        "in": "query",
                        "description": "The id of the unicorn to link to",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "examples": {
                            "application/json": {
                                "id": 1,
                                "user": {
                                    "id": 1,
                                    "firstName": "David",
                                    "lastName": "",
                                    "email": ""
                                },
                                "unicorn": null,
                                "text": "content of the post ...",
                                "createdAt": {
                                    "timezone": {
                                        "name": "Europe/Brussels",
                                        "location": {
                                            "country_code": "BE",
                                            "latitude": 50.83333,
                                            "longitude": 4.33333,
                                            "comments": ""
                                        }
                                    },
                                    "offset": 3600,
                                    "timestamp": 1548204515
                                },
                                "modifiedAt": {
                                    "timezone": {
                                        "name": "Europe/Brussels",
                                        "location": {
                                            "country_code": "BE",
                                            "latitude": 50.83333,
                                            "longitude": 4.33333,
                                            "comments": ""
                                        }
                                    },
                                    "offset": 3600,
                                    "timestamp": 1548204515
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post/update": {
            "post": {
                "tags": [
                    "post"
                ],
                "summary": "Update the content of an existing post.",
                "description": "Update the content of an existing post",
                "operationId": "updatePostAction",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "postId",
                        "in": "query",
                        "description": "The id of the post",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "The content of the post",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "examples": {
                            "application/json": {
                                "id": 1,
                                "user": {
                                    "id": 1,
                                    "firstName": "David",
                                    "lastName": "",
                                    "email": ""
                                },
                                "unicorn": null,
                                "text": "content of the post ...",
                                "createdAt": {
                                    "timezone": {
                                        "name": "Europe/Brussels",
                                        "location": {
                                            "country_code": "BE",
                                            "latitude": 50.83333,
                                            "longitude": 4.33333,
                                            "comments": ""
                                        }
                                    },
                                    "offset": 3600,
                                    "timestamp": 1548204515
                                },
                                "modifiedAt": {
                                    "timezone": {
                                        "name": "Europe/Brussels",
                                        "location": {
                                            "country_code": "BE",
                                            "latitude": 50.83333,
                                            "longitude": 4.33333,
                                            "comments": ""
                                        }
                                    },
                                    "offset": 3600,
                                    "timestamp": 1548204515
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}